Reorganized code into modules
